portal/requirements.txt
django==2.1.7
channels==2.1.7
channels-redis==2.3.2
mysqlclient==1.3.10
django_cryptography==0.3
python-keystoneclient==3.18.0
python-novaclient==11.1.0
python-neutronclient==6.11.0
python-cinderclient==4.1.0
python-glanceclient==2.15.0
celery==4.2.1
redis==3.2.0
gunicorn==19.9.0
service_identity==18.1.0
